requirements.txt
google-adk>=1.5.0
google-genai>=0.3.0
pillow>=11.2.1
python-dotenv>=0.9.9 